management-gui/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
kubernetes>=28.1.0
pyyaml>=6.0
requests>=2.31.0
gunicorn>=21.2.0
